Form container separated from App, fixed re-rendering issue
